+++ CMakeLists.txt
@@ -34,11 +34,14 @@ SET_TARGET_PROPERTIES(ATen PROPERTIES
 
 ### Python Headers for pytorch
 FIND_PACKAGE(PythonLibs REQUIRED)
-TARGET_INCLUDE_DIRECTORIES(autogradpp PUBLIC 
-  ${PYTHON_INCLUDE_DIRS} 
-  "${CMAKE_CURRENT_SOURCE_DIR}"
-  "${CMAKE_CURRENT_SOURCE_DIR}/pytorch"
-  "${CMAKE_CURRENT_SOURCE_DIR}/include"
+TARGET_INCLUDE_DIRECTORIES(autogradpp 
+  PUBLIC 
+    ${PYTHON_INCLUDE_DIRS} 
+    "${CMAKE_CURRENT_SOURCE_DIR}"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pytorch"
+  PRIVATE
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/autogradpp"
 )
 
 ### Copy over all the relevant .so files
@@ -62,6 +65,11 @@ ADD_LIBRARY(libpythpp SHARED IMPORTED)
 SET_TARGET_PROPERTIES(libpythpp PROPERTIES IMPORTED_NO_SONAME 1
   IMPORTED_LOCATION "${CMAKE_BINARY_DIR}/libpythpp.so")
 
+ADD_LIBRARY(libgloo SHARED IMPORTED)
+SET_TARGET_PROPERTIES(libgloo PROPERTIES IMPORTED_NO_SONAME 1
+  IMPORTED_LOCATION "${CMAKE_CURRENT_SOURCE_DIR}/pytorch/torch/lib/libgloo.so")
+
+
 IF(APPLE)
   # Set RPATHs for OSX...
   ADD_CUSTOM_COMMAND(
@@ -74,7 +82,7 @@ IF(APPLE)
 ENDIF(APPLE)
 
 ### Link stuff
-TARGET_LINK_LIBRARIES(autogradpp ATen libpythpp ${PYTHON_LIBRARIES})
+TARGET_LINK_LIBRARIES(autogradpp ATen libpythpp libgloo ${PYTHON_LIBRARIES})
 
 ### Optionally, link CUDA
 FIND_PACKAGE(CUDA)
