+++ src/containers.cpp
@@ -1,4 +1,4 @@
-#include "containers.h"
+#include "autogradpp/containers.h"
 
 namespace autograd {
 std::map<std::string, Variable> ContainerImpl::parameters() const {
